+++ src/main/java/DukeException.java
@@ -11,17 +11,5 @@ public class DukeException extends Exception {
             return false;
         }
     }
-
-    public static int checkTaskIndex(String userInput, int maxIndex) throws DukeException {
-        int index = Integer.parseInt(userInput.trim()) - 1;
-        if (maxIndex < 0){
-            throw new DukeException("Error! There are no task in the list");
-        }
-        if (index > maxIndex){
-            throw new DukeException("Index not found! Choose a valid index!");
-        }
-        return index;
-    }
-
 }
